+++ qudarap/qscint.h
@@ -8,38 +8,38 @@ qscint.h
 #if defined(__CUDACC__) || defined(__CUDABE__)
    #define QSCINT_METHOD __device__
 #else
-   #define QSCINT_METHOD 
-#endif 
+   #define QSCINT_METHOD
+#endif
 
 
 #include "qrng.h"
 
-struct quad4 ; 
-struct quad6 ; 
-struct sphoton ; 
+struct quad4 ;
+struct quad6 ;
+struct sphoton ;
 
 #include "OpticksPhoton.h"
 
 struct qscint
 {
-    cudaTextureObject_t scint_tex ; 
-    quad4*              scint_meta ; // HUH: not used ? 
-    unsigned            hd_factor ; 
+    cudaTextureObject_t scint_tex ;
+    quad4*              scint_meta ; // HUH: not used ?
+    unsigned            hd_factor ;
 
 #if defined(__CUDACC__) || defined(__CUDABE__) || defined(MOCK_CURAND) || defined(MOCK_CUDA)
-    QSCINT_METHOD void    generate( sphoton& p, RNG& rng, const quad6& gs, int photon_id, int genstep_id ) const ; 
+    QSCINT_METHOD void    generate( sphoton& p, RNG& rng, const quad6& gs, unsigned long long photon_id, int genstep_id ) const ;
     QSCINT_METHOD void    reemit(   sphoton& p, RNG& rng, float scintillationTime) const ;
-    QSCINT_METHOD void    momw_polw(sphoton& p, RNG& rng) const ; 
+    QSCINT_METHOD void    momw_polw(sphoton& p, RNG& rng) const ;
     // sets direction, polarization and wavelength as needed by both generate and reemit
 
-    QSCINT_METHOD float   wavelength(     const float& u0) const ; 
-    QSCINT_METHOD float   wavelength_hd0( const float& u0) const ;  
+    QSCINT_METHOD float   wavelength(     const float& u0) const ;
+    QSCINT_METHOD float   wavelength_hd0( const float& u0) const ;
     QSCINT_METHOD float   wavelength_hd10(const float& u0) const ;
     QSCINT_METHOD float   wavelength_hd20(const float& u0) const ;
 
 #endif
 
-}; 
+};
 
 
 #if defined(__CUDACC__) || defined(__CUDABE__) || defined(MOCK_CURAND) || defined(MOCK_CUDA)
@@ -53,24 +53,24 @@ qscint::generate_photon
 **/
 
 inline QSCINT_METHOD void qscint::generate(
-    sphoton& p, 
-    RNG& rng, 
-    const quad6& _gs, 
-    int photon_id, 
-    int genstep_id ) const 
+    sphoton& p,
+    RNG& rng,
+    const quad6& _gs,
+    unsigned long long photon_id,
+    int genstep_id ) const
 {
-    momw_polw(p, rng ); 
+    momw_polw(p, rng );
 
-    const sscint& gs = (const sscint&)_gs ; 
+    const sscint& gs = (const sscint&)_gs ;
 
-    float fraction = gs.charge == 0.f  ? 1.f : curand_uniform(&rng) ;   
-    p.pos = gs.pos + fraction*gs.DeltaPosition ; 
+    float fraction = gs.charge == 0.f  ? 1.f : curand_uniform(&rng) ;
+    p.pos = gs.pos + fraction*gs.DeltaPosition ;
 
-    float u4 = curand_uniform(&rng) ; 
+    float u4 = curand_uniform(&rng) ;
     float deltaTime = fraction*gs.step_length/gs.meanVelocity - gs.ScintillationTime*logf(u4) ;
 
-    p.time = gs.time + deltaTime ; 
-    p.zero_flags(); 
+    p.time = gs.time + deltaTime ;
+    p.zero_flags();
     p.set_flag(SCINTILLATION) ;
 }
 
@@ -81,27 +81,27 @@ qscint::reemit_photon
 
 OLD NOTES IN NEED OF REVIST : HOW TO HANDLE REEMISSION scintillationTime ?
 
-As reemission happens inside scintillators for photons arising from Cerenkov (or Torch) 
+As reemission happens inside scintillators for photons arising from Cerenkov (or Torch)
 gensteps need to special case the handing of the reemission scintillationTime somehow
-because do not have access to scintillation gensteps when handling cerenkov or torch photons. 
+because do not have access to scintillation gensteps when handling cerenkov or torch photons.
 
-Could carry the single float (could be domain compressed, it is eg 1.5 ns) in other gensteps ? 
-But it is material specific (if you had more than one scintillator) 
-just like REEMISSIONPROB so its more appropriate 
+Could carry the single float (could be domain compressed, it is eg 1.5 ns) in other gensteps ?
+But it is material specific (if you had more than one scintillator)
+just like REEMISSIONPROB so its more appropriate
 to live in the boundary_tex alongside the REEMISSIONPROB ?
 
 But it could be carried in the genstep(or anywhere) as its use is "gated" by a non-zero REEMISSIONPROB.
 
-Prefer to just hold it in the context, and provide G4Opticks::setReemissionScintillationTime API 
-for setting it (default 0.) that is used from detector specific code which can read from 
+Prefer to just hold it in the context, and provide G4Opticks::setReemissionScintillationTime API
+for setting it (default 0.) that is used from detector specific code which can read from
 the Geant4 properties directly.  What about geocache ? Can hold/persist with GScintillatorLib metadata.
 
 **/
 
-inline QSCINT_METHOD void qscint::reemit(sphoton& p, RNG& rng, float scintillationTime) const 
+inline QSCINT_METHOD void qscint::reemit(sphoton& p, RNG& rng, float scintillationTime) const
 {
-    momw_polw(p, rng); 
-    float u3 = curand_uniform(&rng) ; 
+    momw_polw(p, rng);
+    float u3 = curand_uniform(&rng) ;
     p.time += -scintillationTime*logf(u3) ;
 }
 
@@ -114,12 +114,12 @@ Translation of "jcv DsG4Scintillation"
 
 **/
 
-inline QSCINT_METHOD void qscint::momw_polw(sphoton& p, RNG& rng) const 
+inline QSCINT_METHOD void qscint::momw_polw(sphoton& p, RNG& rng) const
 {
-    float u0 = curand_uniform(&rng); 
-    float u1 = curand_uniform(&rng); 
-    float u2 = curand_uniform(&rng); 
-    float u3 = curand_uniform(&rng); 
+    float u0 = curand_uniform(&rng);
+    float u1 = curand_uniform(&rng);
+    float u2 = curand_uniform(&rng);
+    float u3 = curand_uniform(&rng);
 
     float cost = 1.f - 2.f*u0;
     float sint = sqrt((1.f-cost)*(1.f+cost));
@@ -127,45 +127,45 @@ inline QSCINT_METHOD void qscint::momw_polw(sphoton& p, RNG& rng) const
     float sinp = sin(phi);
     float cosp = cos(phi);
 
-    p.mom.x = sint*cosp;  
+    p.mom.x = sint*cosp;
     p.mom.y = sint*sinp;
-    p.mom.z = cost ;  
-    p.orient_iindex = 0u ; 
+    p.mom.z = cost ;
+    p.orient_iindex = 0u ;
 
-    // Determine polarization of new photon 
-    p.pol.x = cost*cosp ; 
-    p.pol.y = cost*sinp ; 
+    // Determine polarization of new photon
+    p.pol.x = cost*cosp ;
+    p.pol.y = cost*sinp ;
     p.pol.z = -sint ;
 
     phi = 2.f*M_PIf*u2 ;
-    sinp = sin(phi); 
-    cosp = cos(phi); 
+    sinp = sin(phi);
+    cosp = cos(phi);
 
-    p.pol = normalize( cosp*p.pol + sinp*cross(p.mom, p.pol) ) ;   
+    p.pol = normalize( cosp*p.pol + sinp*cross(p.mom, p.pol) ) ;
     p.wavelength = wavelength(u3);
 }
 
 
 
-inline QSCINT_METHOD float qscint::wavelength(const float& u0) const 
+inline QSCINT_METHOD float qscint::wavelength(const float& u0) const
 {
-    float wl ;  
+    float wl ;
     switch(hd_factor)
-    {   
-        case 0:  wl = wavelength_hd0(u0)  ; break ; 
-        case 10: wl = wavelength_hd10(u0) ; break ; 
-        case 20: wl = wavelength_hd20(u0) ; break ; 
-        default: wl = 0.f ; 
-    }   
-    //printf("//qscint::wavelength wl %10.4f hd %d \n", wl, hd_factor ); 
-    return wl ; 
+    {
+        case 0:  wl = wavelength_hd0(u0)  ; break ;
+        case 10: wl = wavelength_hd10(u0) ; break ;
+        case 20: wl = wavelength_hd20(u0) ; break ;
+        default: wl = 0.f ;
+    }
+    //printf("//qscint::wavelength wl %10.4f hd %d \n", wl, hd_factor );
+    return wl ;
 }
 
 
-inline QSCINT_METHOD float qscint::wavelength_hd0(const float& u0) const 
+inline QSCINT_METHOD float qscint::wavelength_hd0(const float& u0) const
 {
-    constexpr float y0 = 0.5f/3.f ; 
-    return tex2D<float>(scint_tex, u0, y0 ); 
+    constexpr float y0 = 0.5f/3.f ;
+    return tex2D<float>(scint_tex, u0, y0 );
 }
 
 /**
@@ -180,52 +180,52 @@ icdf texture can share some of teh implementation
 
 **/
 
-inline QSCINT_METHOD float qscint::wavelength_hd10(const float& u0) const 
+inline QSCINT_METHOD float qscint::wavelength_hd10(const float& u0) const
 {
-    float wl ; 
+    float wl ;
 
-    constexpr float y0 = 0.5f/3.f ; 
-    constexpr float y1 = 1.5f/3.f ; 
-    constexpr float y2 = 2.5f/3.f ; 
+    constexpr float y0 = 0.5f/3.f ;
+    constexpr float y1 = 1.5f/3.f ;
+    constexpr float y2 = 2.5f/3.f ;
 
     if( u0 < 0.1f )
     {
-        wl = tex2D<float>(scint_tex, u0*10.f , y1 );    
+        wl = tex2D<float>(scint_tex, u0*10.f , y1 );
     }
     else if ( u0 > 0.9f )
     {
-        wl = tex2D<float>(scint_tex, (u0 - 0.9f)*10.f , y2 );    
+        wl = tex2D<float>(scint_tex, (u0 - 0.9f)*10.f , y2 );
     }
     else
     {
-        wl = tex2D<float>(scint_tex, u0,  y0 ); 
+        wl = tex2D<float>(scint_tex, u0,  y0 );
     }
-    return wl ; 
+    return wl ;
 }
 
 
 
-inline QSCINT_METHOD float qscint::wavelength_hd20(const float& u0) const 
+inline QSCINT_METHOD float qscint::wavelength_hd20(const float& u0) const
 {
-    float wl ; 
+    float wl ;
 
-    constexpr float y0 = 0.5f/3.f ; 
-    constexpr float y1 = 1.5f/3.f ; 
-    constexpr float y2 = 2.5f/3.f ; 
+    constexpr float y0 = 0.5f/3.f ;
+    constexpr float y1 = 1.5f/3.f ;
+    constexpr float y2 = 2.5f/3.f ;
 
     if( u0 < 0.05f )
     {
-        wl = tex2D<float>(scint_tex, u0*20.f , y1 );    
+        wl = tex2D<float>(scint_tex, u0*20.f , y1 );
     }
     else if ( u0 > 0.95f )
     {
-        wl = tex2D<float>(scint_tex, (u0 - 0.95f)*20.f , y2 );    
+        wl = tex2D<float>(scint_tex, (u0 - 0.95f)*20.f , y2 );
     }
     else
     {
-        wl = tex2D<float>(scint_tex, u0,  y0 ); 
+        wl = tex2D<float>(scint_tex, u0,  y0 );
     }
-    return wl ; 
+    return wl ;
 }
